Extract BatchUploadModifiedCoordinates.ExportTask
